# docs/swagger.yaml
basePath: /v1
definitions:
  http.DefaultResponse-array_model_CinemaReservation:
    properties:
      data:
        items:
          $ref: '#/definitions/model.CinemaReservation'
        type: array
      error_message:
        type: string
      status:
        type: string
    type: object
  http.DefaultResponse-error:
    properties:
      data: {}
      error_message:
        type: string
      status:
        type: string
    type: object
  http.DefaultResponse-model_Apartment:
    properties:
      data:
        $ref: '#/definitions/model.Apartment'
      error_message:
        type: string
      status:
        type: string
    type: object
  http.DefaultResponse-string:
    properties:
      data:
        type: string
      error_message:
        type: string
      status:
        type: string
    type: object
  http.bindApartmentRequest:
    properties:
      apartment_id:
        type: string
    required:
    - apartment_id
    type: object
  http.confirmRequest:
    properties:
      otp_code:
        type: string
      username:
        type: string
    required:
    - otp_code
    - username
    type: object
  http.createApartmentRequest:
    properties:
      door_num:
        type: integer
      floor:
        type: integer
    required:
    - door_num
    - floor
    type: object
  http.createReservationRequest:
    properties:
      from:
        type: string
      people_num:
        type: integer
      to:
        type: string
    required:
    - from
    - people_num
    - to
    type: object
  http.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  http.registerRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - first_name
    - last_name
    - password
    - username
    type: object
  model.Apartment:
    properties:
      door_number:
        type: integer
      floor:
        type: integer
      id:
        type: string
      owner_id:
        type: string
    type: object
  model.CinemaReservation:
    properties:
      end_time:
        type: string
      id:
        type: string
      is_approved:
        type: boolean
      people_num:
        type: integer
      start_time:
        type: string
      user_id:
        type: string
    type: object
host: assyl-c9b2197f0ace.herokuapp.com
info:
  contact: {}
  description: API для работы с приложением для ЖК.
  title: Assyl Backend API
  version: "1.0"
paths:
  /apartment:
    get:
      consumes:
      - application/json
      description: Возвращает квартиру по этажу и номеру двери.
      parameters:
      - description: Этаж
        in: query
        name: floor
        required: true
        type: integer
      - description: Номер двери
        in: query
        name: door_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Квартира найдена
          schema:
            $ref: '#/definitions/http.DefaultResponse-model_Apartment'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "404":
          description: Квартира не найдена
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      security:
      - JWT: []
      summary: Get apartment
      tags:
      - apartment
  /apartment/bind:
    post:
      consumes:
      - application/json
      description: Привязывает квартиру к текущему авторизованному пользователю.
      parameters:
      - description: Bind apartment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.bindApartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Квартира успешно привязана к пользователю
          schema:
            $ref: '#/definitions/http.DefaultResponse-string'
        "400":
          description: Невалидный запрос (невалидный JSON или тело)
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "401":
          description: Неавторизован (отсутствует или неверный JWT)
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "404":
          description: Квартира не найдена
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "409":
          description: Квартира уже привязана к пользователю
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      security:
      - JWT: []
      summary: Bind apartment to user
      tags:
      - apartment
  /apartment/create:
    post:
      consumes:
      - application/json
      description: Создаёт новую квартиру.
      parameters:
      - description: Create apartment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createApartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Квартира успешно создана
          schema:
            $ref: '#/definitions/http.DefaultResponse-string'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "409":
          description: Квартира уже существует (конфликт)
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      security:
      - JWT: []
      summary: Create apartment
      tags:
      - apartment
  /auth/confirm:
    post:
      consumes:
      - application/json
      description: Подтверждает пользователя по OTP-коду.
      parameters:
      - description: Confirm request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.confirmRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Подтверждение прошло успешно
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      summary: Confirm registration
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Заходит по логину отдавая jwt токен.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успех
          schema:
            $ref: '#/definitions/http.DefaultResponse-string'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      summary: Login into account
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя и отправляет OTP на указанный
        username (телефон/почта).
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/http.DefaultResponse-string'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      summary: Register new user
      tags:
      - auth
  /reservation:
    get:
      consumes:
      - application/json
      description: Возвращает список бронирований текущего пользователя за период
        (from-to).
      parameters:
      - description: Start datetime (RFC3339)
        example: "2025-12-29T10:00:00Z"
        in: query
        name: from
        required: true
        type: string
      - description: End datetime (RFC3339)
        example: "2025-12-29T12:00:00Z"
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успех
          schema:
            $ref: '#/definitions/http.DefaultResponse-array_model_CinemaReservation'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      security:
      - BearerAuth: []
      summary: Get user reservations
      tags:
      - reservation
    post:
      consumes:
      - application/json
      description: Создаёт бронирование на указанный период для текущего пользователя.
      parameters:
      - description: Create reservation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успех
          schema:
            $ref: '#/definitions/http.DefaultResponse-string'
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      security:
      - BearerAuth: []
      summary: Create reservation
      tags:
      - reservation
  /reservation/approve:
    patch:
      description: Одобряет бронирование по reservation_id. Доступно только администратору.
      parameters:
      - description: reservation_id
        in: query
        name: reservation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Одобрено
        "400":
          description: Невалидный запрос
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "403":
          description: Только админ может одобрять
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/http.DefaultResponse-error'
      security:
      - BearerAuth: []
      summary: Approve reservation
      tags:
      - reservation
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"